backend_easyhealth/requirements.txt
asgiref==3.8.1
build==1.0.3
CacheControl==0.13.1
cachetools==6.2.2
certifi==2025.7.14
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
crashtest==0.4.1
crispy-bootstrap5==2023.10
cryptography==42.0.1
defusedxml==0.7.1
distlib==0.3.7
dj-database-url==2.1.0
Django==5.0.2
django-3-jet==1.0.8
django-4-jet==1.0.9
django-admin-argon-dashboard==1.0.16
django-cors-headers==4.2.0
django-crispy-forms==2.1
django-environ==0.11.2
django-unfold
django-jet==1.0.8
django-material-admin==1.8.6
django-redis==5.4.0
django-registration-redux==2.13
django-shopping-cart==0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dulwich==0.21.7
EasyProcess==1.1
entrypoint2==1.1
fastjsonschema==2.19.1
filelock==3.13.1
requests==2.31.0
idna==3.6
importlib-metadata==7.0.1
installer==0.7.0
jaraco.classes==3.3.1
keyring==24.3.0
MarkupSafe==2.1.4
more-itertools==10.2.0
MouseInfo==0.1.3
msgpack==1.0.7
mss==9.0.1
numpy==1.26.2
oauth2client==4.1.3
oauthlib==3.2.2
opencv-contrib-python==4.9.0.80
packaging==23.2
pandas==2.1.3
pexpect==4.9.0
phonenumbers==9.0.9
pillow==11.0.0
pipenv==2023.11.17
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycountry==24.6.1
pycparser==2.21
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
pyproject_hooks==1.0.0
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.30
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
pytweening==1.0.7
pytz==2023.3.post1
rapidfuzz==3.6.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.9
shellingham==1.5.4
six==1.16.0
social-auth-core==4.5.1
sqlparse==0.5.3
stripe==8.3.0
tomlkit==0.12.3
toutatis==1.31
trove-classifiers==2024.1.31
typing_extensions==4.9.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.24.7
xattr==0.10.1
zipp==3.17.0
google-genai>=1.0.0

